requirements.txt
ollama
python-dotenv
ddgs
httpx
questionary
python-docx
reportlab
arabic-reshaper
python-bidi